requirements.txt
jiwer==3.0.5
numpy==2.1.3
openai==1.57.0
openai_whisper==20240930
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==11.0.0
Requests==2.32.3
torch==2.4.0
tqdm==4.66.5
transformers==4.46.3
vllm==0.5.5
